% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcularCapacidad.R
\name{calcularCapacidad}
\alias{calcularCapacidad}
\title{CalcularCapacidad}
\usage{
calcularCapacidad(rutaStr = NULL, nodos = NULL, demandas = NULL,
  maxCap = NULL)
}
\arguments{
\item{rutaStr}{Lista con las rutas calculadas por CW hasta el momento}

\item{nodos}{Vector con los nodos por los que se van a unir las rutas (edge[i,j])}

\item{demandas}{Vector con las demandas de los clientes. Mismo orden que los nombres}

\item{maxCap}{Capacidad máxima del vehículo}
}
\value{
cupo: Boolean. TRUE si la restricción de capacidad de cumple, 0 dlc

capacidad: Capacidad de la ruta luego de unir.
}
\description{
Evalua la restricción de capacidad de la heurística CW
}
