% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWCVRP.R
\name{CWVRP}
\alias{CWVRP}
\title{Clarke Wright Capacity Constrained Vehicle Routing Problem}
\usage{
CWVRP(maxCap = NULL, dist = NULL, clients = LETTERS[1:(dim(dist)[1] -
  1)], dem = NULL)
}
\arguments{
\item{maxCap}{: Capacidad máxima del vehiculo}

\item{depo:}{Data frame que incluya las siguientes variables: id del nodo,Coordenada x, Coordenada y, Demanda. Solamente utilizará la primera fila (1 depósito).}

\item{client:}{Data frame que incluya las siguientes variables: id del nodo,Coordenada x, Coordenada y, Demanda}
}
\value{
Devuelve una lista de rutas asignadas mediante la heurística CW
}
\description{
Clarke Wright Capacity Constrained Vehicle Routing Problem
}
\examples{
dem <- c(10,15,18,17,3,5,9,4,6)
dist <- matrix(c(0,12,11,7,10,10,9,8,6,12,12,0,8,5,9,12,14,16,17,22,11,8,0,9,15,17,8,18,14,22,7,5,9,0,7,9,11,12,12,17,10,9,15,7,0,3,17,7,15,18,10,12,17,9,3,0,18,6,15,15,9,14,8,11,17,18,0,16,8,16,8,16,18,12,7,6,16,0,11,11,6,17,14,12,15,15,8,11,0,10,12,22,22,17,18,15,16,11,10,0),ncol=10,nrow=10)
CWVRP(maxCap = 40,dist = dist,clients = 1:9,dem = dem)
}
