% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMat.R
\name{distMat}
\alias{distMat}
\title{distMat}
\usage{
distMat(coordMat = NULL)
}
\arguments{
\item{coordMat}{data frame con las coordenadas x y y de los nodos}
}
\description{
Calcula la matriz de distancias a partir de coordenadas X y Y
}
