% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eucdist.R
\name{euc.dist}
\alias{euc.dist}
\title{eucDist}
\usage{
euc.dist(x1, y1, x2, y2)
}
\arguments{
\item{x1}{}

\item{y1}{}

\item{x2}{}

\item{y2}{}
}
\description{
Calcula las distancias euclidianas a partir de las coordenadas X y Y
}
