% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrizAhorros.R
\name{matrizAhorros}
\alias{matrizAhorros}
\title{matrizAhorros}
\usage{
matrizAhorros(distMat = NULL)
}
\arguments{
\item{distMat}{Matriz de distancias. Distancia del nodo i al nodo j. Ver distMat}
}
\value{
Matriz de Ahorros
}
\description{
Calcula la matriz de ahorros a partir de la matriz de distancias
}
