% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodoAsignado.R
\name{nodoAsignado}
\alias{nodoAsignado}
\title{Nodo Asignado}
\usage{
nodoAsignado(rutaStr, nodos)
}
\arguments{
\item{rutaStr}{Lista con las rutas calculadas por CW hasta el momento}

\item{nodos}{Vector con los nodos por los que se van a unir las rutas (edge[i,j])}
}
\value{
Boolen: True si se cumple, False de lo contrario
}
\description{
Evalúa la restricción de CW de que los nodos deben estar en rutas diferentes
}
