% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pegarRuta.R
\name{pegarRuta}
\alias{pegarRuta}
\title{Pegar Ruta}
\usage{
pegarRuta(nodos = NULL, rutaStr = NULL, noms = NULL)
}
\arguments{
\item{nodos}{vector con los nodos por los cuales se uniran las rutas (edge[i,j]). Nota:length(nodos)==2}

\item{rutaStr}{Lista con las rutas hasta el momento desarrolladas por la heuristica CW}

\item{noms}{(optional)Vector que contiene los nombres de los clientes}
}
\description{
Function que pega las rutas una vez se evaluó la factibilidad par ala heurística de Clarke-Wright. Ver CWCVRP
}
