% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primeroUltimo.R
\name{primeroUltimo}
\alias{primeroUltimo}
\title{TPrimeroUltimo}
\usage{
primeroUltimo(nodos = NULL, rutaStr = NULL)
}
\arguments{
\item{nodos}{Vector con los nodos por los que se van a unir las rutas (edge[i,j])}

\item{rutaStr}{Lista con las rutas calculadas por CW hasta el momento}
}
\description{
Resulelve si un par de nodos cumplen la restriccion de CW. Uno de ellos debe estar al inicio de alguna la ruta y el otro al final
}
