% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorAhorros.R
\name{vectorAhorros}
\alias{vectorAhorros}
\title{vectorAhorros}
\usage{
vectorAhorros(ahorros = NULL, noms = LETTERS[1:dim(ahorros)[1]])
}
\arguments{
\item{ahorros}{Matriz de ahorros calculada con la función matrizAhorros}

\item{noms}{Vector con las nombres de los clientes. Mismo orden que los nodos de la matriz}
}
\value{
data frame con [nodo i, nodo j, ahorro]
}
\description{
Convierte la matriz de ahorros en el formato deseado [i,j,ahorro]
}
